pyproject.toml
[project]
name = "docker-sentinel"
version = "0.13.0"
requires-python = ">=3.11"
dependencies = [
    "google-adk>=1.0.0",
    "litellm>=1.63.11",
    "anthropic>=0.43.0",
    "docker>=7.0.0",
    "requests>=2.31.0",
    "click>=8.1.0",
    "rich>=13.7.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.6.0",
    "pydantic-settings>=2.2.0",
]

[project.scripts]
docker-sentinel = "docker_sentinel.cli:main"

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"
